Refactor: rename BEM classes to y scheduler
