+++ src/index.html
@@ -23,9 +23,9 @@
               <div class="date-item">
                 <div class="date-item__wrapper"></div>
               </div>
-              <div class="scheduler_users-wrapper">
-                <div class="scheduler_users"></div>
-                <div class="scheduler_todos"></div>
+              <div class="scheduler__executor">
+                <div class="executor-item"></div>
+                <div class="scheduler__todos"></div>
               </div>
             </div>
           </div>
