+++ src/modules/backlog/backlog.js
@@ -1,4 +1,4 @@
-import { dragElem } from '../../classes/Drap_and_Drop'
+import { dragElem } from '../../classes/Drag_and_Drop'
 import listFilter from '../../classes/Inputs'
 import { TASKS } from '../../constants/api/api_urls_scheduler'
 
@@ -10,9 +10,9 @@ const renderTodos = async () => {
   const arr = await [...data]
 
   const wrapperBacklog = document.querySelector('.backlog__list')
-  const wrapperUser = document.querySelectorAll('.todos_user')
+  const wrapperUser = document.querySelectorAll('.scheduler__row')
   const wrapperDate = document.querySelectorAll('.date-item__current')
-  const widthTodos = document.querySelector('.todos').offsetHeight
+  const widthTodos = document.querySelector('.scheduler__list').offsetHeight
 
   arr.forEach((el) => {
     if (el.executor === null) {
@@ -25,7 +25,7 @@ const renderTodos = async () => {
       todoContent.innerText = el.subject
       wrapperBacklog.append(todo)
       todo.addEventListener('dragstart', (e) => {
-        dragElem(e, { textPopup: `Выполнить за: ${el.executionTime}ч`, class: 'backlog__todo', classPopup: 'todo_popup' })
+        dragElem(e, { textPopup: `Выполнить за: ${el.executionTime}ч`, class: 'backlog__todo', classPopup: 'scheduler__popup' })
       })
     } else {
       const dataDate = el.planStartDate
@@ -46,10 +46,10 @@ const renderTodos = async () => {
             todoUser.append(todoContent)
             todoUser.append(todoPopup)
             todoUser.draggable = true
-            todoPopup.classList.add('todo_popup')
+            todoPopup.classList.add('scheduler__popup')
             todoPopup.innerText = `Выполнить за: ${el.executionTime}ч`
             todoUser.style.height = `${(widthTodos / 8) * el.executionTime}px`
-            todoUser.classList.add('user_todo')
+            todoUser.classList.add('scheduler__todo')
             todoContent.innerText = el.subject
 
             for (let k = 0; k < wrapperTodos.length; k++) {
