+++ src/modules/sheulder_users/sheulder_users.js
@@ -1,9 +1,9 @@
-import { allowDropElem, dragElem, dropElem } from '../../classes/Drap_and_Drop'
+import { allowDropElem, dragElem, dropElem } from '../../classes/Drag_and_Drop'
 import { USERS } from '../../constants/api/api_urls_scheduler'
 
 const renderUsers = async () => {
-  const usersWrapper = document.querySelector('.scheduler_users')
-  const todosWrapper = document.querySelector('.scheduler_todos')
+  const usersWrapper = document.querySelector('.executor-item')
+  const todosWrapper = document.querySelector('.scheduler__todos')
 
   const res = await fetch(USERS)
   const data = await res.json()
@@ -13,27 +13,32 @@ const renderUsers = async () => {
 
   arrUsers.forEach((user) => {
     const elemTodos = document.createElement('div')
-    elemTodos.classList.add('todos_user')
+    elemTodos.classList.add('scheduler__row')
 
     dateList.forEach(() => {
       const elem = document.createElement('ul')
-      elem.classList.add('todos')
-      elem.id = 'todos'
+      elem.classList.add('scheduler__list')
       elemTodos.append(elem)
 
       elem.ondragover = allowDropElem
       elem.addEventListener('drop', (e) => {
-        dropElem(e, { class: 'user_todo', classRemove: 'backlog__todo', height: elem.offsetHeight, parent: 'todos', children: 'user_todo' })
+        dropElem(e, {
+          class: 'scheduler__todo',
+          classRemove: 'backlog__todo',
+          height: elem.offsetHeight,
+          parent: 'scheduler__list',
+          children: 'scheduler__todo',
+        })
       })
       elem.addEventListener('drag', (e) => {
-        dragElem(e, { class: 'user_todo' })
+        dragElem(e, { class: 'scheduler__todo' })
       })
     })
 
     const elemUser = document.createElement('span')
     elemUser.innerText = user.name
     usersWrapper.append(elemUser)
-    elemUser.classList.add('user')
+    elemUser.classList.add('executor-item__current')
     todosWrapper.append(elemTodos)
   })
 }
